first version of cat folding and changed to Long for values and update bootcamp

+++ src/main/scala/Fill.scala
@@ -5,7 +5,7 @@ import Fill._;
 import IOdir._;
 
 object Fill {
-  def fillWidthOf(i: Int, n: Node) = { (m: Node) => m.inputs(i).width * n.maxNum }
+  def fillWidthOf(i: Int, n: Node) = { (m: Node) => (m.inputs(i).width * n.maxNum).toInt }
   def apply(n: Int, mod: Bits): Bits = {
     val fillcell = new FillCell(mod);
     fillcell.io.mod <> mod;
